move Phase-2 Tracker records into their own files
