+++ CondFormats/DataRecord/src/SiPhase2OuterTrackerCondDataRecords.cc
@@ -1,0 +1,5 @@
+#include "FWCore/Framework/interface/eventsetuprecord_registration_macro.h"
+#include "CondFormats/DataRecord/interface/SiPhase2OuterTrackerCondDataRecords.h"
+
+EVENTSETUP_RECORD_REG(SiPhase2OuterTrackerLorentzAngleRcd);
+EVENTSETUP_RECORD_REG(SiPhase2OuterTrackerLorentzAngleSimRcd);
