+++ CondFormats/DataRecord/src/SiStripCondDataRecords.cc
@@ -42,5 +42,3 @@ EVENTSETUP_RECORD_REG(SiStripApvSimulationParametersRcd);
 
 EVENTSETUP_RECORD_REG(Phase2TrackerCablingRcd);
 
-EVENTSETUP_RECORD_REG(SiPhase2OuterTrackerLorentzAngleRcd);
-EVENTSETUP_RECORD_REG(SiPhase2OuterTrackerLorentzAngleSimRcd);
